// src/frontend/src/entities/schedule/requests/models/Schedule.ts
/*
 *  Copyright (C) <2022>  <becon GmbH>
 *
 *  This program is free software: you can redistribute it and/or modify
 *  it under the terms of the GNU General Public License as published by
 *  the Free Software Foundation, version 3 of the License.
 *  This program is distributed in the hope that it will be useful,
 *  but WITHOUT ANY WARRANTY; without even the implied warranty of
 *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
 *  GNU General Public License for more details.
 *
 *  You should have received a copy of the GNU General Public License
 *  along with this program. If not, see <http://www.gnu.org/licenses/>.
 */

import ModelWebhook from "./Webhook";

//api description of schedule
export default interface ModelSchedule{
    cronExp: string;
    debugMode: boolean;
    schedulerId?: number;
    status: number;
    title: string;
    connectionId: number;
    connection?: any;
    lastExecution?: any;
    notification?: any[];
    webhook?: ModelWebhook;
}